+++ lib/datagrid/filters/date_filter.rb
@@ -12,7 +12,7 @@ module Datagrid
       end
 
       def apply(grid_object, scope, value)
-        if !dummy? && grid_object.driver.timestamp_column?(scope, name)
+        if !dummy? && !block && grid_object.driver.timestamp_column?(scope, name)
           value = Datagrid::Utils.format_date_as_timestamp(value)
         end
         super