+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [2.0.8]
 
+* Prevent typecasting of date into timestamp when `:date` filter has a block
 * Rescue StatementInvalid when guessing column type from a query
 
 ## [2.0.7]